task_2/requirements.txt
numpy==1.26.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.2.2
pillow==10.0.1
matplotlib==3.7.2
